Feature - Load data in App component, add props to WordCloud component, onclick open twitter page in new window
